Rename button-icons to buttons
