+++ src/config.js
@@ -76,7 +76,6 @@ export default {
   SortableList: false,
   GraphViewerAvailable: true,
   MathAvailable: (typeof MathJax !== 'undefined'),
-  EditorAvailable: true,
   EditorEnabled: false,
   ContentEditable: false,
   WebExtension: ((window.chrome && chrome.runtime && chrome.runtime.id) || (typeof browser !== 'undefined' && browser.runtime && browser.runtime.id)),
