+++ src/editor/toolbar/author/author.js
@@ -3,7 +3,7 @@ import { wrapInList, liftListItem } from "prosemirror-schema-list"
 import { DOMSerializer, DOMParser } from "prosemirror-model"
 import { TextSelection } from "prosemirror-state"
 import { schema, allowedEmptyAttributes } from "./../../schema/base.js"
-import { getButtonHTML } from "./../../../ui/button-icons.js"
+import { getButtonHTML } from "../../../ui/buttons.js"
 import { formHandlerA, formHandlerAnnotate, formHandlerBlockquote, formHandlerImg, formHandlerQ, formHandlerCitation, formHandlerSemantics } from "./handlers.js"
 import { ToolbarView, annotateFormControls } from "../toolbar.js"
 import { escapeCharacters, getCitationOptionsHTML, getLanguageOptionsHTML } from "../../../doc.js"
